Fixing server component
